+++ src/ska_dlm_client/__init__.py
@@ -10,7 +10,7 @@ from ska_dlm_client.utils import CmdLineParameters
 
 __author__ = """Mark Boulton"""
 __email__ = "mark.boulton@uwa.edu.au"
-__version__ = "1.2.2"
+__version__ = "1.2.3"
 
 DLM_CLIENT_LIB_DIR = Path(__file__).parent
 """The library install path of dlm-client."""